refactored code to use helper/test_helper
